Added support for trips without a trip ID
A number of invoices are missing their respective
trip itineraries and reservation forms, and so we
do not have a trip ID # to tie trips to.

This changes allows an "X" prefix and hyphenated
numbers for trip IDs, so that we can have unique
identifiers for trips.

The format will be Xyyyymmdd where yyyy is the year,
mm is the zero leading month, and dd is the zero
leading day. If an ID already exists, a hyphen and
number will be appended (ex. X20231129-2).