Added a tooltip to cost stats on the passenger page
Added a tooltip to cost stats on the passenger page
since it may not be immediately clear what the
difference between "Passenger Cost" and "Total Trip
Cost" is.

Passenger Cost is the cost that specific passenger
has accrued relative to the cost-per-passenger of
each trip they have been on.

Total Trip Cost is the total cost of all trips the
passenger has been on, not respecitve of the
cost-per-passenger.

Maybe these costs shouldn't be displayed together
and it should be visualized in another way ... idk
I am not an expert at UX..